requirements.txt
pandas
openpyxl
requests
beautifulsoup4
google-genai
duckduckgo_search
python-dotenv
xlrd
